feat: Add getProductsByPage function
This commit adds a new file `get-products-by-page.ts` which contains the implementation of the `getProductsByPage` function. This function retrieves products from the API based on the specified page and limit parameters. It makes use of the `tesloApi` instance to send a GET request to the `/products` endpoint with the appropriate query parameters. The response data is then mapped to the `Product` entity using the `ProductMapper` class. If an error occurs during the API request, an error is thrown with the message "Error fetching products".